PDFBOX-2147: Removed unnecessary unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1613617 13f79535-47bb-0310-9956-ffa450edef68

+++ fontbox/src/main/java/org/apache/fontbox/ttf/CMAPEncodingEntry.java
@@ -545,7 +545,7 @@ public class CMAPEncodingEntry
     public int getGlyphId(int characterCode)
     {
         Integer glyphId = characterCodeToGlyphId.get(characterCode);
-        return glyphId == null ? 0 : glyphId.intValue();
+        return glyphId == null ? 0 : glyphId;
     }
 
     /**
